asset map

+++ src/main/java/com/valuequo/buckswise/domain/Assetmapping.java
@@ -23,7 +23,7 @@ public class Assetmapping implements Serializable {
     private Long id;
 
     @Column(name = "uid")
-    private Integer uid;
+    private Long uid;
 
     @Column(name = "goalid")
     private Integer goalid;
@@ -66,16 +66,16 @@ public class Assetmapping implements Serializable {
         this.id = id;
     }
 
-    public Integer getUid() {
+    public Long getUid() {
         return uid;
     }
 
-    public Assetmapping uid(Integer uid) {
+    public Assetmapping uid(Long uid) {
         this.uid = uid;
         return this;
     }
 
-    public void setUid(Integer uid) {
+    public void setUid(Long uid) {
         this.uid = uid;
     }
